Symlink public assets from plugins into `public`.
This allows Dashboard plugins to include CSS and Javascript files that can
easily become a part of the more integrated site.  To take advantage of this,
plugin developers should create a `public` directory in the top level of their
plugin, and populate it with static assets.

Conflicts between names of static assets in Dashboard and in plugins are not
handled gracefully.

Paired-With: Max Martin